% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessStatus.R
\name{getProcessStatus}
\alias{getProcessStatus}
\title{Return process status information}
\usage{
getProcessStatus(psPath = "/bin/ps")
}
\arguments{
\item{psPath}{Full path of the \code{ps} command.}
}
\value{
A dataframe of information from the \code{ps} command.
}
\description{
Returns a dataframe of information from the \code{ps} command.
}
\examples{
\dontrun{
library(dplyr)

# CPU hogs
cpu_hogs <- getProcessStatus() \%>\%
  select(pcpu,user) \%>\%
  group_by(user) \%>\%
  summarise(cpu_usage = sum(pcpu)) \%>\%
  arrange(desc(cpu_usage))

head(cpu_hogs)

# Long processes
long_running <- ps_DF \%>\%
  select(user, esecs, pid, command) \%>\%
  filter(user != "root") \%>\%
  arrange(desc(esecs))

head(long_running)
}
}
